requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
openai>=1.30.0
anthropic>=0.28.0
google-generativeai>=0.5.0
requests>=2.31.0
tldextract>=5.1.0
google-genai>=1.0.0
